+++ auth/auth_helper_test.go
@@ -13,19 +13,19 @@ type AuthHelperSuite struct {
 
 func (a *AuthHelperSuite) TestNormalizeRegistry() {
 	quay := "https://quay.io/v1/"
-	dock := "https://registry.hub.docker.com/v1/"
+	dockv1 := "https://index.docker.io/v1/"
+	dockv2 := "https://index.docker.io/v2/"
 	a.Equal(quay, NormalizeRegistry("https://quay.io"))
 	a.Equal(quay, NormalizeRegistry("https://quay.io/v1"))
 	a.Equal(quay, NormalizeRegistry("http://quay.io/v1"))
 	a.Equal(quay, NormalizeRegistry("https://quay.io/v1/"))
 	a.Equal(quay, NormalizeRegistry("quay.io"))
 
-	a.Equal(dock, NormalizeRegistry(""))
-	a.Equal(dock, NormalizeRegistry("https://registry.hub.docker.com"))
-	a.Equal(dock, NormalizeRegistry("http://registry.hub.docker.com"))
-	a.Equal(dock, NormalizeRegistry("registry.hub.docker.com"))
+	a.Equal(dockv2, NormalizeRegistry(""))
+	a.Equal(dockv1, NormalizeRegistry("https://index.docker.io"))
+	a.Equal(dockv1, NormalizeRegistry("http://index.docker.io"))
+	a.Equal(dockv1, NormalizeRegistry("index.docker.io"))
 	a.Equal("https://quay.io/v2/", NormalizeRegistry("quay.io/v2/"))
-	a.Equal("https://registry-1.docker.io/v2/", NormalizeRegistry("registry-1.docker.io"))
 }
 
 func TestExampleTestSuite(t *testing.T) {